io: move all WriteObject methods implementation into TBufferIO
